training/requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.196
Pillow==10.0.1
numpy==1.24.3
PyYAML==6.0.1
tqdm==4.66.1 